move swap code from quote-service to non-working bin crate swap-service
do some clean-up of the repository for quote-service and remove swap
    code to a newly created swap-service subcrate.
The swap-service is not working
