+++ Cargo.toml
@@ -2,7 +2,7 @@
 resolver = "2"
 
 
-members = [ "crates/bcr-wdc-webapi", "crates/bcr-wdc-keys", "crates/bcr-wdc-quote-service"]
+members = [ "crates/bcr-wdc-webapi", "crates/bcr-wdc-keys", "crates/bcr-wdc-quote-service", "crates/bcr-wdc-swap-service"]
 
 
 [workspace.dependencies]
