+++ crates/bcr-wdc-quote-service/src/utils.rs
@@ -34,10 +34,8 @@ pub fn calculate_default_expiration_date_for_quote(now: crate::TStamp) -> super:
 pub mod tests {
 
     use super::*;
-    use cashu::dhke as cdk_dhke;
     use cashu::nuts::nut01 as cdk01;
     use cashu::nuts::nut02 as cdk02;
-    use cashu::secret as cdk_secret;
 
     pub const RANDOMS: [&str; 6] = [
         "0244e4420934530b2bdf5161f4c88b3c4f923db158741da51f3bb22b579495862e",
@@ -53,66 +51,6 @@ pub mod tests {
             .map(|key| cdk01::PublicKey::from_hex(key).unwrap())
             .collect()
     }
-
-    pub fn generate_proofs(
-        keyset: &cdk02::MintKeySet,
-        amounts: &[cdk_Amount],
-    ) -> Vec<cdk00::Proof> {
-        let mut proofs: Vec<cdk00::Proof> = Vec::new();
-        for amount in amounts {
-            let keypair = keyset.keys.get(amount).expect("keys for amount");
-            let secret = cdk_secret::Secret::new(rand::random::<u64>().to_string());
-            let (b_, r) =
-                cdk_dhke::blind_message(secret.as_bytes(), None).expect("cdk_dhke::blind_message");
-            let c_ =
-                cdk_dhke::sign_message(&keypair.secret_key, &b_).expect("cdk_dhke::sign_message");
-            let c =
-                cdk_dhke::unblind_message(&c_, &r, &keypair.public_key).expect("unblind_message");
-            proofs.push(cdk00::Proof::new(*amount, keyset.id, secret, c));
-        }
-        proofs
-    }
-
-    pub fn generate_blinds(
-        keyset: &cdk02::MintKeySet,
-        amounts: &[cdk_Amount],
-    ) -> Vec<(cdk00::BlindedMessage, cdk_secret::Secret, cdk01::SecretKey)> {
-        let mut blinds: Vec<(cdk00::BlindedMessage, cdk_secret::Secret, cdk01::SecretKey)> =
-            Vec::new();
-        for amount in amounts {
-            let _keypair = keyset.keys.get(amount).expect("keys for amount");
-            let secret = cdk_secret::Secret::new(rand::random::<u64>().to_string());
-            let (b_, r) =
-                cdk_dhke::blind_message(secret.as_bytes(), None).expect("cdk_dhke::blind_message");
-            blinds.push((
-                cdk00::BlindedMessage::new(*amount, keyset.id, b_),
-                secret,
-                r,
-            ));
-        }
-        blinds
-    }
-
-    pub fn verify_signatures_data(
-        keyset: &cdk02::MintKeySet,
-        signatures: impl std::iter::IntoIterator<Item = (cdk00::BlindedMessage, cdk00::BlindSignature)>,
-    ) -> bool {
-        for signature in signatures.into_iter() {
-            if signature.0.keyset_id != keyset.id || signature.1.keyset_id != keyset.id {
-                return false;
-            }
-            if signature.0.amount != signature.1.amount {
-                return false;
-            }
-
-            let keypair = keyset.keys.get(&signature.0.amount);
-            if keypair.is_none() {
-                return false;
-            }
-        }
-        true
-    }
-
     #[test]
     fn test_select_blind_signatures_no_valid_blinds() {
         let publics = publics();
