+++ crates/bcr-wdc-swap-service/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "bcr-wdc-swap-service"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
